(1) Fixed problems in parallel meshing (2) Fixed an infinite recursion in NOX_TSF_Vector (3) Added unit tests for parallel mesh numbering (4) Cleaned up excessive output in parallel (5) Cleaned up some warning messages (6) Moved python meshing utilities to a new directory

+++ src-core/InternalExprs/SundanceDerivOfSymbFunc.cpp
@@ -86,7 +86,7 @@ Set<MultiSet<int> > DerivOfSymbFunc
     {
       MultiSet<int> f;
       f.put(*j);
-      if (maxOrder < 0 || (maxOrder >= 0 && f.size() <= maxOrder))
+      if (maxOrder < 0 || (maxOrder >= 0 && ((int)f.size()) <= maxOrder))
         {
           rtn.put(f);
         }
@@ -98,7 +98,7 @@ Set<MultiSet<int> > DerivOfSymbFunc
       const MultiSet<int>& d = *i;
       if (d.size() >= 1) continue;
       SUNDANCE_VERB_MEDIUM(tabs << "deriv from outside is " << d.toString());
-      if (maxOrder<0 || (maxOrder >= 0 && d.size() <= maxOrder))
+      if (maxOrder<0 || (maxOrder >= 0 && ((int)d.size()) <= maxOrder))
         {
           rtn.put(d);
         }
